10255: Extracted setting state to new action;
