+++ web-client/src/presenter/actions/setSelectedConsolidatedCasesToMultiDocketOnAction.ts
@@ -2,23 +2,6 @@ import { state } from '@web-client/presenter/app.cerebral';
 
 export const setSelectedConsolidatedCasesToMultiDocketOnAction =
   (casesSelected: boolean) =>
-  ({ get, store }: ActionProps) => {
+  ({ store }: ActionProps) => {
     store.set(state.setSelectedConsolidatedCasesToMultiDocketOn, casesSelected);
-    const selectedOrderCases = get(
-      state.modal.form.consolidatedCasesToMultiDocketOn,
-    );
-    store.set(state.createOrderSelectedCases, selectedOrderCases);
-
-    console.log(
-      'sssss',
-      selectedOrderCases
-        .filter(aCase => aCase.checked)
-        .map(aCase => aCase.docketNumberWithSuffix),
-    );
-    store.set(
-      state.addedDocketNumbers,
-      selectedOrderCases
-        .filter(aCase => aCase.checked)
-        .map(aCase => aCase.docketNumberWithSuffix),
-    );
   };
