+++ web-client/src/presenter/sequences/submitUpdateAddDocketNumbersToOrderSequence.ts
@@ -1,7 +1,9 @@
 import { clearModalAction } from '../actions/clearModalAction';
+import { setCreateOrderSelectedCasesAction } from '@web-client/presenter/actions/setCreateOrderSelectedCasesAction';
 import { setSelectedConsolidatedCasesToMultiDocketOnAction } from '../actions/setSelectedConsolidatedCasesToMultiDocketOnAction';
 
 export const submitUpdateAddDocketNumbersToOrderSequence = [
   clearModalAction,
+  setCreateOrderSelectedCasesAction,
   setSelectedConsolidatedCasesToMultiDocketOnAction(true),
 ];
